Package: move pointer file removal

+++ storage_service/locations/models/package.py
@@ -1147,7 +1147,8 @@ class Package(models.Model):
             else:
                 reingest_pointer = os.path.join(ss_internal.full_path, utils.uuid_to_path(self.uuid), reingest_pointer_name)
 
-                # Remove initial copy of pointer
+            # Remove initial copy of pointer
+            if was_compressed:
                 os.remove(os.path.join(ss_internal.full_path, reingest_pointer_name))
 
         # Replace METS
